refactor : 주문 목록 조회 기능 수정
