+++ gc-coffee/src/main/java/hello/gccoffee/controller/AdminApiController.java
@@ -2,6 +2,7 @@ package hello.gccoffee.controller;
 
 import hello.gccoffee.dto.ProductDTO;
 import hello.gccoffee.exception.ProductTaskException;
+import hello.gccoffee.service.OrderMainService;
 import hello.gccoffee.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -18,6 +19,8 @@ import java.util.Map;
 @RequestMapping("/api/products")
 public class AdminApiController {
     private final ProductService productService;
+    private final OrderMainService orderMainService;
+
 
     @PostMapping
     public ResponseEntity<ProductDTO> registerProduct(@Validated @RequestBody ProductDTO productDTO) {
@@ -83,5 +86,24 @@ public class AdminApiController {
                     "message", "An unexpected error occurred"
             ));
         }
+    }
+
+    @GetMapping("/orderlist")   //관리자 주문 목록 조회
+    public ResponseEntity<?> getList(@RequestParam("adminPassword") String adminPassword) {
+
+        log.info("===== getList() =====");
+
+        //비밀번호 없을 때
+        if (adminPassword == null || adminPassword.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Admin password is required");
+        }
+
+        //비밀번호 틀렸을 때
+        if (!"1111".equals(adminPassword)){
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body("Incorrect admin password.");
+        }
+        return ResponseEntity.ok(orderMainService.getList());
     }
 }
